+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `searchdomain` to append a domain suffix to the Proxmox hostname when LINSTOR nodes are registered with FQDNs
 - Support for APIVER14
 - support for LINSTOR auth tokens
+- Support for APIVER15
 
 ## 8.2.0 - 2026-02-24
 ### Added