+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## 8.3.0 - 2026-06-29
 ### Added
+- allow local activate
 - `clusterlocktimeout` for flocks on parallel actions
 - `searchdomain` to append a domain suffix to the Proxmox hostname when LINSTOR nodes are registered with FQDNs
 - Support for APIVER14